Split native UI bridge implementation
